Update test for testing model extract checksums methods
